Fixed scheduled job 'party_is_soon', updated tables, improved some outputs

+++ bot/db/models/wishlist_user_association.py
@@ -1,10 +1,19 @@
-from sqlalchemy import Column, Table, Integer, ForeignKey
+from sqlalchemy import Column, ForeignKey, BigInteger, Table
+from sqlalchemy.orm import relationship
 
 from bot.db.base import Base
+
 
 wishlist_user_association = Table(
     'wishlist_user_association',
     Base.metadata,
-    Column('user_id', Integer, ForeignKey('users.id', ondelete="CASCADE")),
-    Column('wishlist_id', Integer, ForeignKey('wishlists.id', ondelete="CASCADE"))
+    Column('user_id', BigInteger, ForeignKey('users.id', ondelete="CASCADE")),
+    Column('wishlist_id', BigInteger, ForeignKey('wishlists.id', ondelete="CASCADE"))
 )
+
+
+# class WishlistUserAssociation(Base):
+#     __tablename__ = 'wishlist_user_association'
+#
+#     user_id = Column(BigInteger, ForeignKey('users.id', ondelete="CASCADE"), primary_key=True),
+#     wishlist_id = Column(BigInteger, ForeignKey('wishlists.id', ondelete="CASCADE"), primary_key=True)
